link variables
add the ability to specify token/value substitutions in link
urls. They can be defined in the site or panel config.
