+++ provisioning/dashboardData/linksSite.yaml
@@ -1,0 +1,6 @@
+---
+
+linkVariables:
+  urlSiteWebsiteBase: "https://github.com/andymchugh/andrewbmchugh-flow-panel"
+  urlSiteWebsiteSurname: "mchugh"
+  animationSiteVar: "animation"
